Split grocery_item_model.dart
