+++ lib/models/grocery_item_model.dart
@@ -1,110 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-
-class ListItemModel {
-  String id;
-  int? index;
-  bool? isSelected;
-  int? quantity;
-  double? salePricePerItem;
-  Timestamp? dateAddedToList;
-
-  ListItemModel({
-    required this.id,
-    this.index,
-    this.isSelected,
-    this.quantity,
-    this.salePricePerItem,
-    this.dateAddedToList,
-  });
-
-  // serialization
-  factory ListItemModel.fromMap(Map<String, dynamic> data, String documentId) {
-    String id = data['id'];
-    int? quantity = data['quantity'];
-    double? salePricePerItem = data['salePrice'];
-    bool? isSelected = data['isSelected'];
-    int? index = data['index'];
-    Timestamp? dateAdded = data['dateAddedToList'];
-
-    return ListItemModel(
-      id: id,
-      quantity: quantity,
-      salePricePerItem: salePricePerItem,
-      isSelected: isSelected,
-      index: index,
-      dateAddedToList: dateAdded,
-    );
-  }
-
-  Map<String, dynamic> toMap() {
-    return {
-      'id': id,
-      'quantity': quantity,
-      'salePricePerItem': salePricePerItem,
-      'isSelected': isSelected,
-      'index': index,
-      'dateAddedToList': dateAddedToList
-    }..removeWhere((key, value) => value == null);
-  }
-}
-
-class UserItemModel {
-  String id;
-  String? name;
-  String? brand;
-  String? category;
-  double? costPerItem;
-  bool? isFavourite;
-  Timestamp? dateAdded;
-  Timestamp? lastUpdated;
-
-  UserItemModel({
-    required this.id,
-    this.name,
-    this.brand,
-    this.category,
-    this.costPerItem,
-    this.isFavourite,
-    this.dateAdded,
-    this.lastUpdated,
-  });
-
-  // serialization
-  factory UserItemModel.fromMap(Map<String, dynamic> data, String documentId) {
-    String id = documentId;
-    String? name = data['name'];
-    String? brand = data['brand'];
-    double? costPerItem = data['cost'];
-    String? category = data['category'];
-    bool? isFavourite = data['isFavourite'];
-    Timestamp? lastUpdated = data['lastUpdated'];
-    Timestamp? dateAdded = data['dateAdded'];
-
-    return UserItemModel(
-      id: id,
-      name: name,
-      brand: brand,
-      costPerItem: costPerItem,
-      isFavourite: isFavourite,
-      category: category,
-      lastUpdated: lastUpdated,
-      dateAdded: dateAdded,
-    );
-  }
-
-  Map<String, dynamic> toMap() {
-    return {
-      'id': id,
-      'name': name,
-      'brand': brand,
-      'costPerItem': costPerItem,
-      'category': category,
-      'isFavourite': isFavourite,
-      'lastUpdated': lastUpdated,
-      'dateAdded': dateAdded
-    }..removeWhere((key, value) => value == null);
-  }
-}
+import 'package:grocery_list/models/list_item_model.dart';
+import 'package:grocery_list/models/user_item_model.dart';
 
 class GroceryItemModel {
   final ListItemModel groceryListItem;
